+++ packages/frontend/src/components/Terminal.tsx
