+++ packages/frontend/src/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import App from "./App";
+import App from "./components/App";
 
 ReactDOM.render(<App />, document.getElementById("root"));
